adding application.ts and some fe routes
