+++ api/server.ts
@@ -1,38 +1,10 @@
-import * as bodyParser from 'body-parser'
-import * as cookieParser from 'cookie-parser'
 import * as ejs from 'ejs'
 import * as express from 'express'
-import * as session from 'express-session'
 import * as path from 'path'
 import * as process from 'process'
-import * as sessionFileStore from 'session-file-store'
-import * as auth from './auth'
+import { app } from './application'
 import { appInsights } from './lib/appInsights'
-import { environmentConfig } from './lib/environment.config'
-import { errorStack } from './lib/errorStack'
-import * as tunnel from './lib/tunnel'
-import routes from './routes'
 
-const FileStore = sessionFileStore(session)
-
-const app = express()
-
-app.use(
-    session({
-        cookie: {
-            httpOnly: true,
-            maxAge: 1800000,
-            secure: environmentConfig.secureCookie !== false,
-        },
-        name: 'xuiaowebapp',
-        resave: true,
-        saveUninitialized: true,
-        secret: environmentConfig.sessionSecret,
-        store: new FileStore({
-            path: environmentConfig.now ? '/tmp/sessions' : '.sessions',
-        }),
-    })
-)
 
 app.engine('html', ejs.renderFile)
 app.set('view engine', 'html')
@@ -41,20 +13,7 @@ app.set('views', __dirname)
 app.use(express.static(path.join(__dirname, '..', 'assets'), { index: false }))
 app.use(express.static(path.join(__dirname, '..'), { index: false }))
 
-tunnel.init()
-
-app.use(errorStack)
 app.use(appInsights)
-app.use(bodyParser.json())
-app.use(bodyParser.urlencoded({ extended: true }))
-app.use(cookieParser())
-
-app.get('/oauth2/callback', auth.oauth)
-app.get('/api/logout', (req, res, next) => {
-    auth.doLogout(req, res)
-})
-
-app.use('/api', routes)
 
 app.use('/*', (req, res) => {
     console.time(`GET: ${req.originalUrl}`)
