+++ package.json
@@ -44,6 +44,7 @@
     "@angular/platform-browser-dynamic": "^6.0.3",
     "@angular/router": "^6.0.3",
     "@hmcts/frontend": "0.0.39-alpha",
+    "@hmcts/nodejs-healthcheck": "^1.5.1",
     "@hmcts/rpx-xui-common-lib": "0.0.29",
     "@ng-idle/core": "^8.0.0-beta.4",
     "@ng-idle/keepalive": "^8.0.0-beta.4",
