+++ proxy.config.json
@@ -8,5 +8,10 @@
         "target": "http://localhost:3001",
         "secure": false,
         "logLevel": "debug"
+    },
+    "/auth/*": {
+        "target": "http://localhost:3001",
+        "secure": false,
+        "logLevel": "debug"
     }
 }
