+++ src/app/components/index.ts
@@ -7,6 +7,7 @@ import { PrivacyPolicyComponent } from './privacy-policy/privacy-policy.componen
 import { TermsAndConditionsComponent } from './terms-and-conditions/terms-and-conditions.component';
 import { CookiePolicyComponent } from './cookie-policy/cookie-policy.component';
 import { AccessibilityComponent } from './accessibility/accessibility.component';
+import {SignedOutComponent} from './singed-out/signed-out.component';
 
 export const components: any[] = [
   HmctsGlobalHeaderComponent,
@@ -16,7 +17,8 @@ export const components: any[] = [
   CookiePolicyComponent,
   TermsAndConditionsComponent,
   PrivacyPolicyComponent,
-  AccessibilityComponent
+  AccessibilityComponent,
+  SignedOutComponent
 ];
 
 export * from './hmcts-global-header/hmcts-global-header.component';
@@ -28,3 +30,4 @@ export * from './cookie-policy/cookie-policy.component';
 export * from './terms-and-conditions/terms-and-conditions.component';
 export * from './privacy-policy/privacy-policy.component';
 export * from './accessibility/accessibility.component';
+export * from './singed-out/signed-out.component';
