+++ src/app/components/singed-out/signed-out.component.html
@@ -1,7 +1,9 @@
 
-<app-hmcts-main-wrapper [title]="'We have signed you out'">
+<div class="hmcts-width-container">
+  <main class="govuk-main-wrapper">
   <p>For your security, we've signed you out of your account</p>
   <a [href]="redirectUrl" class="govuk-button">Sign in</a>
-</app-hmcts-main-wrapper>
+  </main>
+</div>
