+++ src/app/components/singed-out/signed-out.component.ts
@@ -16,7 +16,4 @@ export class SignedOutComponent implements OnInit {
     this.redirectUrl = './';
   }
 
-
-
-
 }
